completed percentage cut report

+++ Bank.cs
@@ -4,7 +4,7 @@ namespace HeistII_Oceans5
 {
     public class Bank
     {
-        public int CashOnHand { get; set; }
+        public double CashOnHand { get; set; }
         public int AlarmScore { get; set; }
         public int VaultScore { get; set; }
         public int SecurityGuardScore { get; set; }
